pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "ledsa"
version = "0.9.4"
description = "A scientific package to analyse smoke via the dimming of light sources."
authors = [
    {name = "Kristian Börger", email =  "boerger@uni-wuppertal.de"},
    {name = "Lukas Arnold", email =  "arnold@uni-wuppertal.de"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "numpy ~= 1.26.4",
    "matplotlib >= 3.7.2",
    "pandas >= 2.1.2",
    "scipy >= 1.12.0",
    "tables >= 3.9.2",
    "newrawpy >= 0.18.1; platform_machine == 'arm64'",
    "rawpy >= 0.18.1; platform_machine != 'arm64'",
    "tqdm >=4.66.2",
    "exiv2 >= 0.18.0",
    "piexif >= 1.1.3",
    "robotframework >= 6.1.1",
    "pillow >= 10.2.0",
    "opencv-python >= 4.8.0",
    "requests >= 2.31.0",
    "sphinx >= 7.2.6",
    "sphinx-rtd-theme >= 2.0.0",
]

[project.urls]
"Homepage" = "https://firedynamics.github.io/LEDSmokeAnalysis/"
"Repository" = "https://github.com/FireDynamics/LEDSmokeAnalysis"
"Bug Tracker" = "https://github.com/FireDynamics/LEDSmokeAnalysis/issues"
